backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
openai>=1.0.0
pix2tex[gui]
easyocr
paddlepaddle
paddleocr
pytesseract
sympy==1.12
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
